Separate swift modules into files
