+++ compiler/core/src/main.bs.js
@@ -1,18 +1,20 @@
 // Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
 'use strict';
 
-var Fs                         = require("fs");
-var List                       = require("bs-platform/lib/js/list.js");
-var Glob                       = require("glob");
-var Path                       = require("path");
-var $$Array                    = require("bs-platform/lib/js/array.js");
-var Process                    = require("process");
-var FsExtra                    = require("fs-extra");
-var Caml_array                 = require("bs-platform/lib/js/caml_array.js");
-var Color$LonaCompilerCore     = require("./color.bs.js");
-var Swift$LonaCompilerCore     = require("./swift.bs.js");
-var Render$LonaCompilerCore    = require("./render.bs.js");
-var Component$LonaCompilerCore = require("./component.bs.js");
+var Fs                              = require("fs");
+var List                            = require("bs-platform/lib/js/list.js");
+var Glob                            = require("glob");
+var Path                            = require("path");
+var $$Array                         = require("bs-platform/lib/js/array.js");
+var Process                         = require("process");
+var FsExtra                         = require("fs-extra");
+var Caml_array                      = require("bs-platform/lib/js/caml_array.js");
+var Color$LonaCompilerCore          = require("./color.bs.js");
+var Render$LonaCompilerCore         = require("./render.bs.js");
+var Component$LonaCompilerCore      = require("./component.bs.js");
+var SwiftColor$LonaCompilerCore     = require("./swiftColor.bs.js");
+var SwiftRender$LonaCompilerCore    = require("./swiftRender.bs.js");
+var SwiftComponent$LonaCompilerCore = require("./swiftComponent.bs.js");
 
 function exit(message) {
   console.log(message);
@@ -81,7 +83,7 @@ function getTargetExtension(param) {
 var targetExtension = getTargetExtension(target);
 
 function convertColors(filename) {
-  return Swift$LonaCompilerCore.Color[/* render */0](target, Color$LonaCompilerCore.parseFile(filename));
+  return SwiftColor$LonaCompilerCore.render(target, Color$LonaCompilerCore.parseFile(filename));
 }
 
 function convertComponent(filename) {
@@ -92,7 +94,7 @@ function convertComponent(filename) {
     var match = findWorkspaceDirectory(filename);
     if (match) {
       var colors = Color$LonaCompilerCore.parseFile(Path.join(match[0], "colors.json"));
-      return Swift$LonaCompilerCore.Render[/* toString */9](Swift$LonaCompilerCore.Component[/* generate */0](name, parsed, colors));
+      return SwiftRender$LonaCompilerCore.toString(SwiftComponent$LonaCompilerCore.generate(name, parsed, colors));
     } else {
       console.log("Couldn't find workspace directory. Try specifying it as a parameter (TODO)");
       return (process.exit());
@@ -108,7 +110,7 @@ function convertWorkspace(workspace, output) {
   FsExtra.ensureDirSync(toDirectory);
   var colorsInputPath = Path.join(fromDirectory, "colors.json");
   var colorsOutputPath = Path.join(toDirectory, "colors" + targetExtension);
-  var colors = Swift$LonaCompilerCore.Color[/* render */0](target, Color$LonaCompilerCore.parseFile(colorsInputPath));
+  var colors = SwiftColor$LonaCompilerCore.render(target, Color$LonaCompilerCore.parseFile(colorsInputPath));
   Fs.writeFileSync(colorsOutputPath, colors);
   Glob(Path.join(fromDirectory, "**/*.component"), (function (_, files) {
           var files$1 = $$Array.to_list(files);
@@ -135,7 +137,8 @@ switch (command) {
         console.log("No filename given");
         ((process.exit()));
       }
-      console.log(convertColors(Caml_array.caml_array_get(Process.argv, 4)));
+      var filename = Caml_array.caml_array_get(Process.argv, 4);
+      console.log(SwiftColor$LonaCompilerCore.render(target, Color$LonaCompilerCore.parseFile(filename)));
       break;
   case "component" : 
       if (Process.argv.length < 5) {
