+++ compiler/core/src/swiftAst.bs.js
@@ -1,0 +1,2 @@
+// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
+/* This output is empty. Its source's type definitions, externals and/or unused code got optimized away. */
