+++ compiler/core/src/swiftColor.bs.js
@@ -1,0 +1,54 @@
+// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
+'use strict';
+
+var List                         = require("bs-platform/lib/js/list.js");
+var Block                        = require("bs-platform/lib/js/block.js");
+var SwiftRender$LonaCompilerCore = require("./swiftRender.bs.js");
+
+function render(target, colors) {
+  if (target !== 0) {
+    var colorConstantDoc = function (color) {
+      return /* LineEndComment */Block.__(16, [{
+                  comment: color[/* value */2],
+                  line: /* ConstantDeclaration */Block.__(6, [{
+                        modifiers: /* :: */[
+                          /* AccessLevelModifier */Block.__(0, [/* PublicModifier */3]),
+                          /* :: */[
+                            /* StaticModifier */11,
+                            /* [] */0
+                          ]
+                        ],
+                        pattern: /* IdentifierPattern */Block.__(0, [{
+                              identifier: color[/* id */0],
+                              annotation: /* None */0
+                            }]),
+                        init: /* Some */[/* LiteralExpression */Block.__(0, [/* Color */Block.__(4, [color[/* value */2]])])]
+                      }])
+                }]);
+    };
+    return SwiftRender$LonaCompilerCore.toString(/* TopLevelDeclaration */Block.__(19, [{
+                    statements: /* :: */[
+                      /* ImportDeclaration */Block.__(10, ["UIKit"]),
+                      /* :: */[
+                        /* Empty */0,
+                        /* :: */[
+                          /* ClassDeclaration */Block.__(4, [{
+                                name: "Colors",
+                                inherits: /* [] */0,
+                                modifier: /* None */0,
+                                isFinal: /* false */0,
+                                body: List.map(colorConstantDoc, colors)
+                              }]),
+                          /* [] */0
+                        ]
+                      ]
+                    ]
+                  }]));
+  } else {
+    console.log("Color generation not supported for target", target);
+    return "error";
+  }
+}
+
+exports.render = render;
+/* SwiftRender-LonaCompilerCore Not a pure module */
