+++ compiler/core/src/swiftComponent.bs.js
@@ -1,0 +1,1171 @@
+// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
+'use strict';
+
+var List                           = require("bs-platform/lib/js/list.js");
+var Block                          = require("bs-platform/lib/js/block.js");
+var Curry                          = require("bs-platform/lib/js/curry.js");
+var Caml_obj                       = require("bs-platform/lib/js/caml_obj.js");
+var Pervasives                     = require("bs-platform/lib/js/pervasives.js");
+var LodashUpperfirst               = require("lodash.upperfirst");
+var Layer$LonaCompilerCore         = require("./layer.bs.js");
+var Logic$LonaCompilerCore         = require("./logic.bs.js");
+var Decode$LonaCompilerCore        = require("./decode.bs.js");
+var StringMap$LonaCompilerCore     = require("./stringMap.bs.js");
+var SwiftLogic$LonaCompilerCore    = require("./swiftLogic.bs.js");
+var SwiftFormat$LonaCompilerCore   = require("./swiftFormat.bs.js");
+var SwiftDocument$LonaCompilerCore = require("./swiftDocument.bs.js");
+
+function generate(name, json, colors) {
+  var rootLayer = Decode$LonaCompilerCore.Component[/* rootLayer */1](json);
+  var nonRootLayers = List.tl(Layer$LonaCompilerCore.flatten(rootLayer));
+  var logic = Decode$LonaCompilerCore.Component[/* logic */2](json);
+  var assignments = Layer$LonaCompilerCore.parameterAssignmentsFromLogic(rootLayer, logic);
+  var parameters = Decode$LonaCompilerCore.Component[/* parameters */0](json);
+  var priorityName = function (param) {
+    if (param !== 0) {
+      return "defaultLow";
+    } else {
+      return "required";
+    }
+  };
+  var typeAnnotationDoc = function (param) {
+    if (param.tag) {
+      return /* TypeName */Block.__(0, [param[0]]);
+    } else {
+      var typeName = param[0];
+      if (typeName === "Boolean") {
+        return /* TypeName */Block.__(0, ["Bool"]);
+      } else {
+        return /* TypeName */Block.__(0, [typeName]);
+      }
+    }
+  };
+  var parameterVariableDoc = function (parameter) {
+    return /* VariableDeclaration */Block.__(7, [{
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PublicModifier */3]),
+                  /* [] */0
+                ],
+                pattern: /* IdentifierPattern */Block.__(0, [{
+                      identifier: parameter[/* name */0],
+                      annotation: /* Some */[typeAnnotationDoc(parameter[/* ltype */1])]
+                    }]),
+                init: /* None */0,
+                block: /* Some */[/* WillSetDidSetBlock */[{
+                      willSet: /* None */0,
+                      didSet: /* Some */[/* :: */[
+                          /* FunctionCallExpression */Block.__(14, [{
+                                name: /* SwiftIdentifier */Block.__(5, ["update"]),
+                                arguments: /* [] */0
+                              }]),
+                          /* [] */0
+                        ]]
+                    }]]
+              }]);
+  };
+  var viewTypeInitDoc = function (param) {
+    switch (param) {
+      case 0 : 
+          return /* SwiftIdentifier */Block.__(5, ["UIView"]);
+      case 1 : 
+          return /* SwiftIdentifier */Block.__(5, ["UILabel"]);
+      case 2 : 
+          return /* SwiftIdentifier */Block.__(5, ["UIImageView"]);
+      case 3 : 
+      case 4 : 
+      case 5 : 
+      case 6 : 
+          return /* SwiftIdentifier */Block.__(5, ["TypeUnknown"]);
+      
+    }
+  };
+  var viewVariableDoc = function (layer) {
+    var match = +(layer[/* typeName */0] === /* Text */1);
+    return /* VariableDeclaration */Block.__(7, [{
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                  /* [] */0
+                ],
+                pattern: /* IdentifierPattern */Block.__(0, [{
+                      identifier: SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]),
+                      annotation: /* None */0
+                    }]),
+                init: /* Some */[/* FunctionCallExpression */Block.__(14, [{
+                        name: viewTypeInitDoc(layer[/* typeName */0]),
+                        arguments: match !== 0 ? /* [] */0 : /* :: */[
+                            /* FunctionCallArgument */Block.__(13, [{
+                                  name: /* Some */[/* SwiftIdentifier */Block.__(5, ["frame"])],
+                                  value: /* SwiftIdentifier */Block.__(5, [".zero"])
+                                }]),
+                            /* [] */0
+                          ]
+                      }])],
+                block: /* None */0
+              }]);
+  };
+  var spacingVariableDoc = function (layer) {
+    var variableName = function (variable) {
+      var match = +(layer === rootLayer);
+      if (match !== 0) {
+        return variable;
+      } else {
+        return SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]) + LodashUpperfirst(variable);
+      }
+    };
+    var match = +(layer === rootLayer);
+    var marginVariables;
+    if (match !== 0) {
+      marginVariables = /* [] */0;
+    } else {
+      var createVariable = function (marginParameter) {
+        return /* VariableDeclaration */Block.__(7, [{
+                    modifiers: /* :: */[
+                      /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                      /* [] */0
+                    ],
+                    pattern: /* IdentifierPattern */Block.__(0, [{
+                          identifier: variableName(marginParameter[/* swiftName */1]),
+                          annotation: /* Some */[/* TypeName */Block.__(0, ["CGFloat"])]
+                        }]),
+                    init: /* Some */[/* LiteralExpression */Block.__(0, [/* FloatingPoint */Block.__(2, [Layer$LonaCompilerCore.getNumberParameter(marginParameter[/* lonaName */0], layer)])])],
+                    block: /* None */0
+                  }]);
+      };
+      marginVariables = List.map(createVariable, /* :: */[
+            /* record */[
+              /* lonaName */"marginTop",
+              /* swiftName */"topMargin"
+            ],
+            /* :: */[
+              /* record */[
+                /* lonaName */"marginRight",
+                /* swiftName */"trailingMargin"
+              ],
+              /* :: */[
+                /* record */[
+                  /* lonaName */"marginBottom",
+                  /* swiftName */"bottomMargin"
+                ],
+                /* :: */[
+                  /* record */[
+                    /* lonaName */"marginLeft",
+                    /* swiftName */"leadingMargin"
+                  ],
+                  /* [] */0
+                ]
+              ]
+            ]
+          ]);
+    }
+    var match$1 = layer[/* children */3];
+    var paddingVariables;
+    if (match$1) {
+      var createVariable$1 = function (paddingParameter) {
+        return /* VariableDeclaration */Block.__(7, [{
+                    modifiers: /* :: */[
+                      /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                      /* [] */0
+                    ],
+                    pattern: /* IdentifierPattern */Block.__(0, [{
+                          identifier: variableName(paddingParameter[/* swiftName */1]),
+                          annotation: /* Some */[/* TypeName */Block.__(0, ["CGFloat"])]
+                        }]),
+                    init: /* Some */[/* LiteralExpression */Block.__(0, [/* FloatingPoint */Block.__(2, [Layer$LonaCompilerCore.getNumberParameter(paddingParameter[/* lonaName */0], layer)])])],
+                    block: /* None */0
+                  }]);
+      };
+      paddingVariables = List.map(createVariable$1, /* :: */[
+            /* record */[
+              /* lonaName */"paddingTop",
+              /* swiftName */"topPadding"
+            ],
+            /* :: */[
+              /* record */[
+                /* lonaName */"paddingRight",
+                /* swiftName */"trailingPadding"
+              ],
+              /* :: */[
+                /* record */[
+                  /* lonaName */"paddingBottom",
+                  /* swiftName */"bottomPadding"
+                ],
+                /* :: */[
+                  /* record */[
+                    /* lonaName */"paddingLeft",
+                    /* swiftName */"leadingPadding"
+                  ],
+                  /* [] */0
+                ]
+              ]
+            ]
+          ]);
+    } else {
+      paddingVariables = /* [] */0;
+    }
+    return Pervasives.$at(marginVariables, paddingVariables);
+  };
+  var initParameterDoc = function (parameter) {
+    return /* Parameter */Block.__(12, [{
+                externalName: /* None */0,
+                localName: parameter[/* name */0],
+                annotation: typeAnnotationDoc(parameter[/* ltype */1]),
+                defaultValue: /* None */0
+              }]);
+  };
+  var initParameterAssignmentDoc = function (parameter) {
+    return /* BinaryExpression */Block.__(2, [{
+                left: /* MemberExpression */Block.__(1, [/* :: */[
+                      /* SwiftIdentifier */Block.__(5, ["self"]),
+                      /* :: */[
+                        /* SwiftIdentifier */Block.__(5, [parameter[/* name */0]]),
+                        /* [] */0
+                      ]
+                    ]]),
+                operator: "=",
+                right: /* SwiftIdentifier */Block.__(5, [parameter[/* name */0]])
+              }]);
+  };
+  var initializerCoderDoc = function () {
+    return /* InitializerDeclaration */Block.__(8, [{
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PublicModifier */3]),
+                  /* :: */[
+                    /* RequiredModifier */10,
+                    /* [] */0
+                  ]
+                ],
+                parameters: /* :: */[
+                  /* Parameter */Block.__(12, [{
+                        externalName: /* Some */["coder"],
+                        localName: "aDecoder",
+                        annotation: /* TypeName */Block.__(0, ["NSCoder"]),
+                        defaultValue: /* None */0
+                      }]),
+                  /* [] */0
+                ],
+                failable: /* Some */["?"],
+                body: /* :: */[
+                  /* FunctionCallExpression */Block.__(14, [{
+                        name: /* SwiftIdentifier */Block.__(5, ["fatalError"]),
+                        arguments: /* :: */[
+                          /* FunctionCallArgument */Block.__(13, [{
+                                name: /* None */0,
+                                value: /* SwiftIdentifier */Block.__(5, ["\"init(coder:) has not been implemented\""])
+                              }]),
+                          /* [] */0
+                        ]
+                      }]),
+                  /* [] */0
+                ]
+              }]);
+  };
+  var initializerDoc = function () {
+    return /* InitializerDeclaration */Block.__(8, [{
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PublicModifier */3]),
+                  /* [] */0
+                ],
+                parameters: List.map(initParameterDoc, parameters),
+                failable: /* None */0,
+                body: SwiftDocument$LonaCompilerCore.joinGroups(/* Empty */0, /* :: */[
+                      List.map(initParameterAssignmentDoc, parameters),
+                      /* :: */[
+                        /* :: */[
+                          /* MemberExpression */Block.__(1, [/* :: */[
+                                /* SwiftIdentifier */Block.__(5, ["super"]),
+                                /* :: */[
+                                  /* FunctionCallExpression */Block.__(14, [{
+                                        name: /* SwiftIdentifier */Block.__(5, ["init"]),
+                                        arguments: /* :: */[
+                                          /* FunctionCallArgument */Block.__(13, [{
+                                                name: /* Some */[/* SwiftIdentifier */Block.__(5, ["frame"])],
+                                                value: /* SwiftIdentifier */Block.__(5, [".zero"])
+                                              }]),
+                                          /* [] */0
+                                        ]
+                                      }]),
+                                  /* [] */0
+                                ]
+                              ]]),
+                          /* [] */0
+                        ],
+                        /* :: */[
+                          /* :: */[
+                            /* FunctionCallExpression */Block.__(14, [{
+                                  name: /* SwiftIdentifier */Block.__(5, ["setUpViews"]),
+                                  arguments: /* [] */0
+                                }]),
+                            /* :: */[
+                              /* FunctionCallExpression */Block.__(14, [{
+                                    name: /* SwiftIdentifier */Block.__(5, ["setUpConstraints"]),
+                                    arguments: /* [] */0
+                                  }]),
+                              /* [] */0
+                            ]
+                          ],
+                          /* :: */[
+                            /* :: */[
+                              /* FunctionCallExpression */Block.__(14, [{
+                                    name: /* SwiftIdentifier */Block.__(5, ["update"]),
+                                    arguments: /* [] */0
+                                  }]),
+                              /* [] */0
+                            ],
+                            /* [] */0
+                          ]
+                        ]
+                      ]
+                    ])
+              }]);
+  };
+  var memberOrSelfExpression = function (firstIdentifier, statements) {
+    if (firstIdentifier === "self") {
+      return /* MemberExpression */Block.__(1, [statements]);
+    } else {
+      return /* MemberExpression */Block.__(1, [Pervasives.$at(/* :: */[
+                      /* SwiftIdentifier */Block.__(5, [firstIdentifier]),
+                      /* [] */0
+                    ], statements)]);
+    }
+  };
+  var parentNameOrSelf = function (parent) {
+    var match = +(parent === rootLayer);
+    if (match !== 0) {
+      return "self";
+    } else {
+      return SwiftFormat$LonaCompilerCore.layerName(parent[/* name */1]);
+    }
+  };
+  var initialLayerValue = function (layer, name) {
+    var match = StringMap$LonaCompilerCore.find_opt(name, layer[/* parameters */2]);
+    if (match) {
+      return SwiftDocument$LonaCompilerCore.lonaValue(colors, match[0]);
+    } else {
+      var param = name;
+      if (param === "backgroundColor") {
+        return /* MemberExpression */Block.__(1, [/* :: */[
+                    /* SwiftIdentifier */Block.__(5, ["UIColor"]),
+                    /* :: */[
+                      /* SwiftIdentifier */Block.__(5, ["clear"]),
+                      /* [] */0
+                    ]
+                  ]]);
+      } else {
+        return /* LiteralExpression */Block.__(0, [/* Integer */Block.__(1, [0])]);
+      }
+    }
+  };
+  var defineInitialLayerValue = function (layer, param) {
+    var name = param[0];
+    var match = initialLayerValue(layer, name);
+    var match$1;
+    var exit = 0;
+    switch (name) {
+      case "borderRadius" : 
+          if (typeof match === "number") {
+            exit = 1;
+          } else if (match.tag) {
+            exit = 1;
+          } else {
+            var tmp = match[0];
+            if (typeof tmp === "number" || tmp.tag !== 2) {
+              exit = 1;
+            } else {
+              match$1 = /* tuple */[
+                memberOrSelfExpression(parentNameOrSelf(layer), /* :: */[
+                      /* SwiftIdentifier */Block.__(5, ["layer"]),
+                      /* :: */[
+                        /* SwiftIdentifier */Block.__(5, ["cornerRadius"]),
+                        /* [] */0
+                      ]
+                    ]),
+                match
+              ];
+            }
+          }
+          break;
+      case "height" : 
+          if (typeof match === "number") {
+            exit = 1;
+          } else if (match.tag) {
+            exit = 1;
+          } else {
+            var tmp$1 = match[0];
+            if (typeof tmp$1 === "number" || tmp$1.tag !== 2) {
+              exit = 1;
+            } else {
+              match$1 = /* tuple */[
+                /* SwiftIdentifier */Block.__(5, [parentNameOrSelf(layer) + "HeightAnchorConstraint?.constant"]),
+                match
+              ];
+            }
+          }
+          break;
+      case "visible" : 
+          if (typeof match === "number") {
+            exit = 1;
+          } else if (match.tag) {
+            exit = 1;
+          } else {
+            var match$2 = match[0];
+            if (typeof match$2 === "number" || match$2.tag) {
+              exit = 1;
+            } else {
+              match$1 = /* tuple */[
+                memberOrSelfExpression(parentNameOrSelf(layer), /* :: */[
+                      /* SwiftIdentifier */Block.__(5, ["isHidden"]),
+                      /* [] */0
+                    ]),
+                /* LiteralExpression */Block.__(0, [/* Boolean */Block.__(0, [1 - match$2[0]])])
+              ];
+            }
+          }
+          break;
+      case "width" : 
+          if (typeof match === "number") {
+            exit = 1;
+          } else if (match.tag) {
+            exit = 1;
+          } else {
+            var tmp$2 = match[0];
+            if (typeof tmp$2 === "number" || tmp$2.tag !== 2) {
+              exit = 1;
+            } else {
+              match$1 = /* tuple */[
+                /* SwiftIdentifier */Block.__(5, [parentNameOrSelf(layer) + "WidthAnchorConstraint?.constant"]),
+                match
+              ];
+            }
+          }
+          break;
+      default:
+        exit = 1;
+    }
+    if (exit === 1) {
+      match$1 = /* tuple */[
+        memberOrSelfExpression(parentNameOrSelf(layer), /* :: */[
+              /* SwiftIdentifier */Block.__(5, [name]),
+              /* [] */0
+            ]),
+        match
+      ];
+    }
+    return /* BinaryExpression */Block.__(2, [{
+                left: match$1[0],
+                operator: "=",
+                right: match$1[1]
+              }]);
+  };
+  var setUpViewsDoc = function (root) {
+    var setUpDefaultsDoc = function () {
+      var filterParameters = function (param) {
+        var name = param[0];
+        if (name !== "image" && name !== "textStyle" && name !== "flexDirection" && name !== "justifyContent" && name !== "alignSelf" && name !== "alignItems" && name !== "flex" && name !== "font" && !name.startsWith("padding") && !name.startsWith("margin") && name !== "height") {
+          return +(name !== "width");
+        } else {
+          return /* false */0;
+        }
+      };
+      var defineInitialLayerValues = function (layer) {
+        return List.map((function (param) {
+                      return defineInitialLayerValue(layer, /* tuple */[
+                                  param[0],
+                                  param[1]
+                                ]);
+                    }), List.filter((function (param) {
+                            var layer$1 = layer;
+                            var param$1 = param;
+                            var match = Layer$LonaCompilerCore.LayerMap[/* find_opt */24](layer$1, assignments);
+                            if (match) {
+                              var match$1 = StringMap$LonaCompilerCore.find_opt(param$1[0], match[0]);
+                              if (match$1) {
+                                return /* false */0;
+                              } else {
+                                return /* true */1;
+                              }
+                            } else {
+                              return /* true */1;
+                            }
+                          }))(List.filter(filterParameters)(Curry._1(StringMap$LonaCompilerCore.bindings, layer[/* parameters */2]))));
+      };
+      return List.concat(List.map(defineInitialLayerValues, Layer$LonaCompilerCore.flatten(rootLayer)));
+    };
+    var addSubviews = function (parent, layer) {
+      if (parent) {
+        return /* :: */[
+                /* FunctionCallExpression */Block.__(14, [{
+                      name: memberOrSelfExpression(parentNameOrSelf(parent[0]), /* :: */[
+                            /* SwiftIdentifier */Block.__(5, ["addSubview"]),
+                            /* [] */0
+                          ]),
+                      arguments: /* :: */[
+                        /* SwiftIdentifier */Block.__(5, [SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1])]),
+                        /* [] */0
+                      ]
+                    }]),
+                /* [] */0
+              ];
+      } else {
+        return /* [] */0;
+      }
+    };
+    return /* FunctionDeclaration */Block.__(9, [{
+                name: "setUpViews",
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                  /* [] */0
+                ],
+                parameters: /* [] */0,
+                body: SwiftDocument$LonaCompilerCore.joinGroups(/* Empty */0, /* :: */[
+                      List.concat(Layer$LonaCompilerCore.flatmapParent(addSubviews, root)),
+                      /* :: */[
+                        setUpDefaultsDoc(/* () */0),
+                        /* [] */0
+                      ]
+                    ])
+              }]);
+  };
+  var getConstraints = function (root) {
+    var setUpContraint = function (layer, anchor1, parent, anchor2, relation, value, suffix) {
+      var match = +(layer === rootLayer);
+      var variableName = (
+        match !== 0 ? anchor1 : SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]) + LodashUpperfirst(anchor1)
+      ) + suffix;
+      var initialValue = /* MemberExpression */Block.__(1, [/* :: */[
+            /* SwiftIdentifier */Block.__(5, [SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1])]),
+            /* :: */[
+              /* SwiftIdentifier */Block.__(5, [anchor1]),
+              /* :: */[
+                /* FunctionCallExpression */Block.__(14, [{
+                      name: /* SwiftIdentifier */Block.__(5, ["constraint"]),
+                      arguments: /* :: */[
+                        /* FunctionCallArgument */Block.__(13, [{
+                              name: /* Some */[/* SwiftIdentifier */Block.__(5, [relation])],
+                              value: memberOrSelfExpression(parentNameOrSelf(parent), /* :: */[
+                                    /* SwiftIdentifier */Block.__(5, [anchor2]),
+                                    /* [] */0
+                                  ])
+                            }]),
+                        /* :: */[
+                          /* FunctionCallArgument */Block.__(13, [{
+                                name: /* Some */[/* SwiftIdentifier */Block.__(5, ["constant"])],
+                                value: value
+                              }]),
+                          /* [] */0
+                        ]
+                      ]
+                    }]),
+                /* [] */0
+              ]
+            ]
+          ]]);
+      return /* record */[
+              /* variableName */variableName,
+              /* initialValue */initialValue,
+              /* priority : Required */0
+            ];
+    };
+    var setUpLessThanOrEqualToContraint = function (layer, anchor1, parent, anchor2, value, suffix) {
+      var match = +(layer === rootLayer);
+      var variableName = (
+        match !== 0 ? anchor1 : SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]) + LodashUpperfirst(anchor1)
+      ) + suffix;
+      var initialValue = /* MemberExpression */Block.__(1, [/* :: */[
+            /* SwiftIdentifier */Block.__(5, [SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1])]),
+            /* :: */[
+              /* SwiftIdentifier */Block.__(5, [anchor1]),
+              /* :: */[
+                /* FunctionCallExpression */Block.__(14, [{
+                      name: /* SwiftIdentifier */Block.__(5, ["constraint"]),
+                      arguments: /* :: */[
+                        /* FunctionCallArgument */Block.__(13, [{
+                              name: /* Some */[/* SwiftIdentifier */Block.__(5, ["lessThanOrEqualTo"])],
+                              value: memberOrSelfExpression(parentNameOrSelf(parent), /* :: */[
+                                    /* SwiftIdentifier */Block.__(5, [anchor2]),
+                                    /* [] */0
+                                  ])
+                            }]),
+                        /* :: */[
+                          /* FunctionCallArgument */Block.__(13, [{
+                                name: /* Some */[/* SwiftIdentifier */Block.__(5, ["constant"])],
+                                value: value
+                              }]),
+                          /* [] */0
+                        ]
+                      ]
+                    }]),
+                /* [] */0
+              ]
+            ]
+          ]]);
+      return /* record */[
+              /* variableName */variableName,
+              /* initialValue */initialValue,
+              /* priority : Low */1
+            ];
+    };
+    var setUpDimensionContraint = function (layer, anchor, constant) {
+      var match = +(layer === rootLayer);
+      var variableName = (
+        match !== 0 ? anchor : SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]) + LodashUpperfirst(anchor)
+      ) + "Constraint";
+      var initialValue = memberOrSelfExpression(parentNameOrSelf(layer), /* :: */[
+            /* SwiftIdentifier */Block.__(5, [anchor]),
+            /* :: */[
+              /* FunctionCallExpression */Block.__(14, [{
+                    name: /* SwiftIdentifier */Block.__(5, ["constraint"]),
+                    arguments: /* :: */[
+                      /* FunctionCallArgument */Block.__(13, [{
+                            name: /* Some */[/* SwiftIdentifier */Block.__(5, ["equalToConstant"])],
+                            value: /* LiteralExpression */Block.__(0, [/* FloatingPoint */Block.__(2, [constant])])
+                          }]),
+                      /* [] */0
+                    ]
+                  }]),
+              /* [] */0
+            ]
+          ]);
+      return /* record */[
+              /* variableName */variableName,
+              /* initialValue */initialValue,
+              /* priority : Required */0
+            ];
+    };
+    var negateNumber = function (expression) {
+      return /* PrefixExpression */Block.__(3, [{
+                  operator: "-",
+                  expression: expression
+                }]);
+    };
+    var constraintConstantExpression = function (layer, variable1, parent, variable2) {
+      var variableName = function (layer, variable) {
+        var match = +(layer === rootLayer);
+        if (match !== 0) {
+          return variable;
+        } else {
+          return SwiftFormat$LonaCompilerCore.layerName(layer[/* name */1]) + LodashUpperfirst(variable);
+        }
+      };
+      return /* BinaryExpression */Block.__(2, [{
+                  left: /* SwiftIdentifier */Block.__(5, [variableName(layer, variable1)]),
+                  operator: "+",
+                  right: /* SwiftIdentifier */Block.__(5, [variableName(parent, variable2)])
+                }]);
+    };
+    var constrainAxes = function (layer) {
+      var direction = Layer$LonaCompilerCore.getFlexDirection(layer);
+      var match = +(direction === "column");
+      var primaryBeforeAnchor = match !== 0 ? "topAnchor" : "leadingAnchor";
+      var match$1 = +(direction === "column");
+      var primaryAfterAnchor = match$1 !== 0 ? "bottomAnchor" : "trailingAnchor";
+      var match$2 = +(direction === "column");
+      var secondaryBeforeAnchor = match$2 !== 0 ? "leadingAnchor" : "topAnchor";
+      var match$3 = +(direction === "column");
+      var secondaryAfterAnchor = match$3 !== 0 ? "trailingAnchor" : "bottomAnchor";
+      var height = Layer$LonaCompilerCore.getNumberParameterOpt("height", layer);
+      var width = Layer$LonaCompilerCore.getNumberParameterOpt("width", layer);
+      var match$4 = +(direction === "column");
+      var primaryDimension = match$4 !== 0 ? "height" : "width";
+      var match$5 = +(direction === "column");
+      var secondaryDimension = match$5 !== 0 ? "width" : "height";
+      var secondaryDimensionAnchor = secondaryDimension + "Anchor";
+      var sizingRules = Layer$LonaCompilerCore.getSizingRules(Layer$LonaCompilerCore.findParent(rootLayer, layer), layer);
+      var match$6 = +(direction === "column");
+      var primarySizingRule = match$6 !== 0 ? sizingRules[/* height */1] : sizingRules[/* width */0];
+      var match$7 = +(direction === "column");
+      var secondarySizingRule = match$7 !== 0 ? sizingRules[/* width */0] : sizingRules[/* height */1];
+      var flexChildren = List.filter((function (child) {
+                return +(Layer$LonaCompilerCore.getNumberParameter("flex", child) === 1.0);
+              }))(layer[/* children */3]);
+      var addConstraints = function (index, child) {
+        var childSizingRules = Layer$LonaCompilerCore.getSizingRules(/* Some */[layer], child);
+        var match = +(direction === "column");
+        var childSecondarySizingRule = match !== 0 ? childSizingRules[/* width */0] : childSizingRules[/* height */1];
+        var firstViewConstraints;
+        if (index !== 0) {
+          firstViewConstraints = /* [] */0;
+        } else {
+          var match$1 = +(direction === "column");
+          var primaryBeforeConstant = match$1 !== 0 ? constraintConstantExpression(layer, "topPadding", child, "topMargin") : constraintConstantExpression(layer, "leadingPadding", child, "leadingMargin");
+          firstViewConstraints = /* :: */[
+            setUpContraint(child, primaryBeforeAnchor, layer, primaryBeforeAnchor, "equalTo", primaryBeforeConstant, "Constraint"),
+            /* [] */0
+          ];
+        }
+        var lastViewConstraints;
+        if (index === (List.length(layer[/* children */3]) - 1 | 0)) {
+          var match$2 = List.length(flexChildren);
+          var needsPrimaryAfterConstraint = typeof primarySizingRule === "number" ? (
+              primarySizingRule !== 0 || match$2 ? /* true */1 : /* false */0
+            ) : (
+              match$2 ? /* true */1 : /* false */0
+            );
+          var match$3 = +(direction === "column");
+          var primaryAfterConstant = match$3 !== 0 ? constraintConstantExpression(layer, "bottomPadding", child, "bottomMargin") : constraintConstantExpression(layer, "trailingPadding", child, "trailingMargin");
+          lastViewConstraints = needsPrimaryAfterConstraint !== 0 ? /* :: */[
+              setUpContraint(child, primaryAfterAnchor, layer, primaryAfterAnchor, "equalTo", negateNumber(primaryAfterConstant), "Constraint"),
+              /* [] */0
+            ] : /* [] */0;
+        } else {
+          lastViewConstraints = /* [] */0;
+        }
+        var middleViewConstraints;
+        if (index !== 0) {
+          var previousLayer = List.nth(layer[/* children */3], index - 1 | 0);
+          var match$4 = +(direction === "column");
+          var betweenConstant = match$4 !== 0 ? constraintConstantExpression(previousLayer, "bottomMargin", child, "topMargin") : constraintConstantExpression(previousLayer, "trailingMargin", child, "leadingMargin");
+          middleViewConstraints = /* :: */[
+            setUpContraint(child, primaryBeforeAnchor, previousLayer, primaryAfterAnchor, "equalTo", betweenConstant, "Constraint"),
+            /* [] */0
+          ];
+        } else {
+          middleViewConstraints = /* [] */0;
+        }
+        var match$5 = +(direction === "column");
+        var secondaryBeforeConstant = match$5 !== 0 ? constraintConstantExpression(layer, "leadingPadding", child, "leadingMargin") : constraintConstantExpression(layer, "topPadding", child, "topMargin");
+        var match$6 = +(direction === "column");
+        var secondaryAfterConstant = match$6 !== 0 ? constraintConstantExpression(layer, "trailingPadding", child, "trailingMargin") : constraintConstantExpression(layer, "bottomPadding", child, "bottomMargin");
+        var secondaryBeforeConstraint = setUpContraint(child, secondaryBeforeAnchor, layer, secondaryBeforeAnchor, "equalTo", secondaryBeforeConstant, "Constraint");
+        var secondaryAfterConstraint;
+        var exit = 0;
+        if (typeof childSecondarySizingRule === "number") {
+          if (childSecondarySizingRule !== 0) {
+            if (typeof secondarySizingRule === "number" && secondarySizingRule === 0) {
+              secondaryAfterConstraint = /* :: */[
+                setUpContraint(child, secondaryAfterAnchor, layer, secondaryAfterAnchor, "lessThanOrEqualTo", negateNumber(secondaryAfterConstant), "Constraint"),
+                /* [] */0
+              ];
+            } else {
+              exit = 1;
+            }
+          } else {
+            secondaryAfterConstraint = /* :: */[
+              setUpContraint(child, secondaryAfterAnchor, layer, secondaryAfterAnchor, "equalTo", negateNumber(secondaryAfterConstant), "Constraint"),
+              /* [] */0
+            ];
+          }
+        } else {
+          secondaryAfterConstraint = /* [] */0;
+        }
+        if (exit === 1) {
+          secondaryAfterConstraint = /* :: */[
+            setUpContraint(child, secondaryAfterAnchor, layer, secondaryAfterAnchor, "equalTo", negateNumber(secondaryAfterConstant), "Constraint"),
+            /* [] */0
+          ];
+        }
+        var fitContentSecondaryConstraint = typeof secondarySizingRule === "number" && secondarySizingRule !== 0 ? /* :: */[
+            setUpLessThanOrEqualToContraint(child, secondaryDimensionAnchor, layer, secondaryDimensionAnchor, negateNumber(/* BinaryExpression */Block.__(2, [{
+                          left: secondaryBeforeConstant,
+                          operator: "+",
+                          right: secondaryAfterConstant
+                        }])), "ParentConstraint"),
+            /* [] */0
+          ] : /* [] */0;
+        return Pervasives.$at(firstViewConstraints, Pervasives.$at(lastViewConstraints, Pervasives.$at(middleViewConstraints, Pervasives.$at(/* :: */[
+                                secondaryBeforeConstraint,
+                                /* [] */0
+                              ], Pervasives.$at(secondaryAfterConstraint, fitContentSecondaryConstraint)))));
+      };
+      var flexChildrenConstraints;
+      if (flexChildren) {
+        var rest = flexChildren[1];
+        var first = flexChildren[0];
+        if (List.length(rest) > 0) {
+          var sameAnchor = primaryDimension + "Anchor";
+          flexChildrenConstraints = List.mapi((function (param, param$1) {
+                  var anchor = sameAnchor;
+                  var index = param;
+                  var layer = param$1;
+                  return setUpContraint(first, anchor, layer, anchor, "equalTo", /* LiteralExpression */Block.__(0, [/* FloatingPoint */Block.__(2, [0.0])]), "SiblingConstraint" + Pervasives.string_of_int(index));
+                }), rest);
+        } else {
+          flexChildrenConstraints = /* [] */0;
+        }
+      } else {
+        flexChildrenConstraints = /* [] */0;
+      }
+      var heightConstraint = height ? /* :: */[
+          setUpDimensionContraint(layer, "heightAnchor", height[0]),
+          /* [] */0
+        ] : /* [] */0;
+      var widthConstraint = width ? /* :: */[
+          setUpDimensionContraint(layer, "widthAnchor", width[0]),
+          /* [] */0
+        ] : /* [] */0;
+      return List.concat(Pervasives.$at(/* :: */[
+                      heightConstraint,
+                      /* :: */[
+                        widthConstraint,
+                        /* [] */0
+                      ]
+                    ], Pervasives.$at(/* :: */[
+                          flexChildrenConstraints,
+                          /* [] */0
+                        ], List.mapi(addConstraints, layer[/* children */3]))));
+    };
+    return List.concat(Layer$LonaCompilerCore.flatmap(constrainAxes, root));
+  };
+  var constraints = getConstraints(rootLayer);
+  var setUpConstraintsDoc = function (root) {
+    var translatesAutoresizingMask = function (layer) {
+      return /* BinaryExpression */Block.__(2, [{
+                  left: memberOrSelfExpression(parentNameOrSelf(layer), /* :: */[
+                        /* SwiftIdentifier */Block.__(5, ["translatesAutoresizingMaskIntoConstraints"]),
+                        /* [] */0
+                      ]),
+                  operator: "=",
+                  right: /* LiteralExpression */Block.__(0, [/* Boolean */Block.__(0, [/* false */0])])
+                }]);
+    };
+    var defineConstraint = function (def) {
+      return /* ConstantDeclaration */Block.__(6, [{
+                  modifiers: /* [] */0,
+                  init: /* Some */[def[/* initialValue */1]],
+                  pattern: /* IdentifierPattern */Block.__(0, [{
+                        identifier: def[/* variableName */0],
+                        annotation: /* None */0
+                      }])
+                }]);
+    };
+    var setConstraintPriority = function (def) {
+      return /* BinaryExpression */Block.__(2, [{
+                  left: /* MemberExpression */Block.__(1, [/* :: */[
+                        /* SwiftIdentifier */Block.__(5, [def[/* variableName */0]]),
+                        /* :: */[
+                          /* SwiftIdentifier */Block.__(5, ["priority"]),
+                          /* [] */0
+                        ]
+                      ]]),
+                  operator: "=",
+                  right: /* MemberExpression */Block.__(1, [/* :: */[
+                        /* SwiftIdentifier */Block.__(5, ["UILayoutPriority"]),
+                        /* :: */[
+                          /* SwiftIdentifier */Block.__(5, [priorityName(def[/* priority */2])]),
+                          /* [] */0
+                        ]
+                      ]])
+                }]);
+    };
+    var activateConstraints = function () {
+      return /* FunctionCallExpression */Block.__(14, [{
+                  name: /* MemberExpression */Block.__(1, [/* :: */[
+                        /* SwiftIdentifier */Block.__(5, ["NSLayoutConstraint"]),
+                        /* :: */[
+                          /* SwiftIdentifier */Block.__(5, ["activate"]),
+                          /* [] */0
+                        ]
+                      ]]),
+                  arguments: /* :: */[
+                    /* FunctionCallArgument */Block.__(13, [{
+                          name: /* None */0,
+                          value: /* LiteralExpression */Block.__(0, [/* Array */Block.__(5, [List.map((function (def) {
+                                          return /* SwiftIdentifier */Block.__(5, [def[/* variableName */0]]);
+                                        }), constraints)])])
+                        }]),
+                    /* [] */0
+                  ]
+                }]);
+    };
+    var assignConstraint = function (def) {
+      return /* BinaryExpression */Block.__(2, [{
+                  left: /* MemberExpression */Block.__(1, [/* :: */[
+                        /* SwiftIdentifier */Block.__(5, ["self"]),
+                        /* :: */[
+                          /* SwiftIdentifier */Block.__(5, [def[/* variableName */0]]),
+                          /* [] */0
+                        ]
+                      ]]),
+                  operator: "=",
+                  right: /* SwiftIdentifier */Block.__(5, [def[/* variableName */0]])
+                }]);
+    };
+    var assignConstraintIdentifier = function (def) {
+      return /* BinaryExpression */Block.__(2, [{
+                  left: /* MemberExpression */Block.__(1, [/* :: */[
+                        /* SwiftIdentifier */Block.__(5, [def[/* variableName */0]]),
+                        /* :: */[
+                          /* SwiftIdentifier */Block.__(5, ["identifier"]),
+                          /* [] */0
+                        ]
+                      ]]),
+                  operator: "=",
+                  right: /* LiteralExpression */Block.__(0, [/* String */Block.__(3, [def[/* variableName */0]])])
+                }]);
+    };
+    return /* FunctionDeclaration */Block.__(9, [{
+                name: "setUpConstraints",
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                  /* [] */0
+                ],
+                parameters: /* [] */0,
+                body: List.concat(/* :: */[
+                      Layer$LonaCompilerCore.flatmap(translatesAutoresizingMask, root),
+                      /* :: */[
+                        /* :: */[
+                          /* Empty */0,
+                          /* [] */0
+                        ],
+                        /* :: */[
+                          List.map(defineConstraint, constraints),
+                          /* :: */[
+                            List.map(setConstraintPriority, List.filter((function (def) {
+                                          return +(def[/* priority */2] === /* Low */1);
+                                        }))(constraints)),
+                            /* :: */[
+                              /* :: */[
+                                /* Empty */0,
+                                /* [] */0
+                              ],
+                              /* :: */[
+                                /* :: */[
+                                  activateConstraints(/* () */0),
+                                  /* [] */0
+                                ],
+                                /* :: */[
+                                  /* :: */[
+                                    /* Empty */0,
+                                    /* [] */0
+                                  ],
+                                  /* :: */[
+                                    List.map(assignConstraint, constraints),
+                                    /* :: */[
+                                      /* :: */[
+                                        /* LineComment */Block.__(15, ["For debugging"]),
+                                        /* [] */0
+                                      ],
+                                      /* :: */[
+                                        List.map(assignConstraintIdentifier, constraints),
+                                        /* [] */0
+                                      ]
+                                    ]
+                                  ]
+                                ]
+                              ]
+                            ]
+                          ]
+                        ]
+                      ]
+                    ])
+              }]);
+  };
+  var updateDoc = function () {
+    var filterParameters = function (param) {
+      var name = param[0];
+      if (name !== "image") {
+        return +(name !== "textStyle");
+      } else {
+        return /* false */0;
+      }
+    };
+    var conditionallyAssigned = Logic$LonaCompilerCore.conditionallyAssignedIdentifiers(logic);
+    var defineInitialLayerValues = function (param) {
+      var layer = param[0];
+      return List.map((function (param) {
+                    return defineInitialLayerValue(layer, param);
+                  }), List.filter((function (param) {
+                          var layer$1 = layer;
+                          var param$1 = param;
+                          var name = param$1[0];
+                          var isAssigned = function (param) {
+                            return Caml_obj.caml_equal(param[1], /* :: */[
+                                        "layers",
+                                        /* :: */[
+                                          layer$1[/* name */1],
+                                          /* :: */[
+                                            name,
+                                            /* [] */0
+                                          ]
+                                        ]
+                                      ]);
+                          };
+                          return Curry._2(Logic$LonaCompilerCore.IdentifierSet[/* exists */15], isAssigned, conditionallyAssigned);
+                        }))(List.filter(filterParameters)(Curry._1(StringMap$LonaCompilerCore.bindings, param[1]))));
+    };
+    return /* FunctionDeclaration */Block.__(9, [{
+                name: "update",
+                modifiers: /* :: */[
+                  /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                  /* [] */0
+                ],
+                parameters: /* [] */0,
+                body: SwiftDocument$LonaCompilerCore.joinGroups(/* Empty */0, /* :: */[
+                      List.concat(List.map(defineInitialLayerValues, Curry._1(Layer$LonaCompilerCore.LayerMap[/* bindings */16], assignments))),
+                      /* :: */[
+                        SwiftLogic$LonaCompilerCore.toSwiftAST(colors, rootLayer, logic),
+                        /* [] */0
+                      ]
+                    ])
+              }]);
+  };
+  return /* TopLevelDeclaration */Block.__(19, [{
+              statements: /* :: */[
+                /* ImportDeclaration */Block.__(10, ["UIKit"]),
+                /* :: */[
+                  /* ImportDeclaration */Block.__(10, ["Foundation"]),
+                  /* :: */[
+                    /* LineComment */Block.__(15, ["MARK: - " + name]),
+                    /* :: */[
+                      /* Empty */0,
+                      /* :: */[
+                        /* ClassDeclaration */Block.__(4, [{
+                              name: name,
+                              inherits: /* :: */[
+                                /* TypeName */Block.__(0, ["UIView"]),
+                                /* [] */0
+                              ],
+                              modifier: /* Some */[/* PublicModifier */3],
+                              isFinal: /* false */0,
+                              body: List.concat(/* :: */[
+                                    /* :: */[
+                                      /* LineComment */Block.__(15, ["MARK: Lifecycle"]),
+                                      /* [] */0
+                                    ],
+                                    /* :: */[
+                                      /* :: */[
+                                        /* Empty */0,
+                                        /* [] */0
+                                      ],
+                                      /* :: */[
+                                        /* :: */[
+                                          initializerDoc(/* () */0),
+                                          /* [] */0
+                                        ],
+                                        /* :: */[
+                                          /* :: */[
+                                            /* Empty */0,
+                                            /* [] */0
+                                          ],
+                                          /* :: */[
+                                            /* :: */[
+                                              initializerCoderDoc(/* () */0),
+                                              /* [] */0
+                                            ],
+                                            /* :: */[
+                                              /* :: */[
+                                                /* LineComment */Block.__(15, ["MARK: Public"]),
+                                                /* [] */0
+                                              ],
+                                              /* :: */[
+                                                /* :: */[
+                                                  /* Empty */0,
+                                                  /* [] */0
+                                                ],
+                                                /* :: */[
+                                                  List.map(parameterVariableDoc, parameters),
+                                                  /* :: */[
+                                                    /* :: */[
+                                                      /* LineComment */Block.__(15, ["MARK: Private"]),
+                                                      /* [] */0
+                                                    ],
+                                                    /* :: */[
+                                                      /* :: */[
+                                                        /* Empty */0,
+                                                        /* [] */0
+                                                      ],
+                                                      /* :: */[
+                                                        List.map(viewVariableDoc, nonRootLayers),
+                                                        /* :: */[
+                                                          /* :: */[
+                                                            /* Empty */0,
+                                                            /* [] */0
+                                                          ],
+                                                          /* :: */[
+                                                            List.concat(Layer$LonaCompilerCore.flatmap(spacingVariableDoc, rootLayer)),
+                                                            /* :: */[
+                                                              /* :: */[
+                                                                /* Empty */0,
+                                                                /* [] */0
+                                                              ],
+                                                              /* :: */[
+                                                                List.map((function (def) {
+                                                                        var variableName = def[/* variableName */0];
+                                                                        return /* VariableDeclaration */Block.__(7, [{
+                                                                                    modifiers: /* :: */[
+                                                                                      /* AccessLevelModifier */Block.__(0, [/* PrivateModifier */0]),
+                                                                                      /* [] */0
+                                                                                    ],
+                                                                                    pattern: /* IdentifierPattern */Block.__(0, [{
+                                                                                          identifier: variableName,
+                                                                                          annotation: /* Some */[/* OptionalType */Block.__(4, [/* TypeName */Block.__(0, ["NSLayoutConstraint"])])]
+                                                                                        }]),
+                                                                                    init: /* None */0,
+                                                                                    block: /* None */0
+                                                                                  }]);
+                                                                      }), constraints),
+                                                                /* :: */[
+                                                                  /* :: */[
+                                                                    /* Empty */0,
+                                                                    /* [] */0
+                                                                  ],
+                                                                  /* :: */[
+                                                                    /* :: */[
+                                                                      setUpViewsDoc(rootLayer),
+                                                                      /* [] */0
+                                                                    ],
+                                                                    /* :: */[
+                                                                      /* :: */[
+                                                                        /* Empty */0,
+                                                                        /* [] */0
+                                                                      ],
+                                                                      /* :: */[
+                                                                        /* :: */[
+                                                                          setUpConstraintsDoc(rootLayer),
+                                                                          /* [] */0
+                                                                        ],
+                                                                        /* :: */[
+                                                                          /* :: */[
+                                                                            /* Empty */0,
+                                                                            /* [] */0
+                                                                          ],
+                                                                          /* :: */[
+                                                                            /* :: */[
+                                                                              updateDoc(/* () */0),
+                                                                              /* [] */0
+                                                                            ],
+                                                                            /* [] */0
+                                                                          ]
+                                                                        ]
+                                                                      ]
+                                                                    ]
+                                                                  ]
+                                                                ]
+                                                              ]
+                                                            ]
+                                                          ]
+                                                        ]
+                                                      ]
+                                                    ]
+                                                  ]
+                                                ]
+                                              ]
+                                            ]
+                                          ]
+                                        ]
+                                      ]
+                                    ]
+                                  ])
+                            }]),
+                        /* [] */0
+                      ]
+                    ]
+                  ]
+                ]
+              ]
+            }]);
+}
+
+var Format = 0;
+
+var Ast = 0;
+
+var Document = 0;
+
+var Render = 0;
+
+exports.Format   = Format;
+exports.Ast      = Ast;
+exports.Document = Document;
+exports.Render   = Render;
+exports.generate = generate;
+/* lodash.upperfirst Not a pure module */
