+++ compiler/core/src/swiftFormat.bs.js
@@ -1,0 +1,11 @@
+// Generated by BUCKLESCRIPT VERSION 2.1.0, PLEASE EDIT WITH CARE
+'use strict';
+
+var LodashCamelcase = require("lodash.camelcase");
+
+function layerName(layerName$1) {
+  return LodashCamelcase(layerName$1) + "View";
+}
+
+exports.layerName = layerName;
+/* lodash.camelcase Not a pure module */
